[Refactor] 프로젝트 내 코드 리팩토링
- Bracnh 패키지 변경 (domain.branch → domain.branch.entity)
- MataJoApplicationTest 내 home 메서드의 DisplayName 내용 추가
